renamed files and components
