+++ src/components/CostWizard/WizardSteps/NodeConfigStep.tsx
@@ -1,12 +1,10 @@
 import React from 'react';
 import { Title, WizardStep } from '@ui5/webcomponents-react';
 import NextStepButton from '../Buttons/NextStepButton';
-import MachineSetup from '../MachineSetup';
+import MachineSetupForm from '../MachineSetupForm';
 import InfoField from '../common/InfoField';
 
-export default function BaseConfigStep() {
-
-
+export default function NodeConfigStep() {
   return (
     <WizardStep selected titleText="Base Configuration">
       <Title wrappingType="Normal" level="H2" size="H2">
@@ -14,7 +12,7 @@ export default function BaseConfigStep() {
       </Title>
       <div className="StepContent">
         <InfoField info="contains 224 GB of storage by default" />
-        <MachineSetup nodeIndex={0} workerNode={false} />
+        <MachineSetupForm nodeIndex={0} workerNode={false} />
       </div>
       <div className="ButtonContainer">
         <NextStepButton />
